refactor(auth-commands): rename the login test_support module to support
`test_support` holds no tests, so the `test_*` prefix — which marks the split
test modules — was misleading. Rename it to `support` and point the test
modules at `super::support`.

Addresses review feedback on pnpm/pnpm#12851.
