+++ pacquet/crates/auth-commands/src/login/test_web_login_scope.rs
@@ -15,7 +15,7 @@ use serde_json::json;
 
 use super::{
     login,
-    test_support::{PromptScript, ReadScript, client, login_fake, opts, written_settings},
+    support::{PromptScript, ReadScript, client, login_fake, opts, written_settings},
 };
 
 #[tokio::test]
